refactor: added resused components in reusable folder
