--- frontend/src/WebDesignLandingPage/WebDesignProjectsPage.tsx
@@ -1,9 +1,0 @@
-
-export const WebDesignProjectsPage = () => {
-  return (
-    <div className='flex flex-row'>
-      <video></video>
-
-    </div>
-  )
-}
